Rename MessageWriter to MessagePacker.
Writers send data to the network (and the file MessagePacker does not do this).
